requirements.txt
pytest~=8.1.1
allure-pytest~=2.13.5
selene>=2.0.0rc9
python-dotenv~=0.20.0
config~=0.5.1
Appium-Python-Client~=4.0.0
requests~=2.31.0
selenium>=4.19.0
